refactor(mm-next): move `components/story/normal/brief` into shared, add props `contentLayout`
