+++ packages/mirror-media-next/components/story/shared/brief.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import DraftRenderBlock from '../shared/draft-renderer-block'
+import DraftRenderBlock from './draft-renderer-block'
 
 /**
  * @typedef {import('../../../type/theme').Theme} Theme
@@ -25,10 +25,11 @@ const BriefContainer = styled.div`
   font-weight: 500;
   font-size: 18px;
   line-height: 36px;
+
   ${({ theme }) => theme.breakpoint.md} {
     font-weight: 400;
     font-size: 19.2px;
-    padding: 19.2px 48.6px 20.8px 28.4px;
+    padding: 24px 32px;
   }
   *,
   *::before,
@@ -38,20 +39,22 @@ const BriefContainer = styled.div`
 `
 
 /**
- *
+ * Component for render brief in `normal` and `premium` story layout
  * @param {Object} props
  * @param {Brief} props.brief
  * @param {String} [props.sectionSlug]
+ * @param { 'normal' | 'wide' | 'photography' | 'premium' } [props.contentLayout]
  * @returns {JSX.Element}
  */
 export default function ArticleBrief({
   brief = { blocks: [], entityMap: {} },
   sectionSlug = '',
+  contentLayout = 'normal',
 }) {
   return (
     <DraftRenderBlock
       rawContentBlock={brief}
-      contentLayout="normal"
+      contentLayout={contentLayout}
       wrapper={(children) => (
         <BriefContainer sectionSlug={sectionSlug}>{children}</BriefContainer>
       )}
